chore: refactor url into constant
