+++ src/constants/ClientUrls.ts
@@ -1,0 +1,6 @@
+export const ClientUrls = {
+  NTUC: 'https://website-api.omni.fairprice.com.sg/api',
+  Giant: 'https://giant.sg/checkout/cart/checkdelivery',
+  ColdStorage: 'https://coldstorage.com.sg/checkout/cart/checkdelivery',
+  ShengSiong: 'https://www.allforyou.sg/Common'
+};
