+++ src/constants/Vendor.ts
@@ -1,0 +1,6 @@
+export const Vendor = {
+  ShengSiong: 'ShengSiong',
+  NTUC: 'NTUC',
+  Giant: 'Giant',
+  ColdStorage: 'ColdStorage'
+};
