--- src/enum/Vendor.ts
@@ -1,6 +1,0 @@
-export enum Vendor {
-  ShengSiong = 'ShengSiong',
-  NTUC = 'NTUC',
-  Giant = 'Giant',
-  ColdStorage = 'ColdStorage'
-}
